refactor: Relocate `useEditorStyles` to the visual editor component
This hook is used solely within this component at this time.
