+++ src/components/visual-editor/use-editor-styles.js
@@ -10,7 +10,7 @@ import { useMemo } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { unlock } from '../lock-unlock';
+import { unlock } from '../../lock-unlock';
 // The Vite query parameter breaks the linter's import resolution
 // eslint-disable-next-line import/no-unresolved
 import defaultThemeStyles from './default-theme-styles.scss?inline';
